+++ config/queue.php
@@ -13,7 +13,8 @@ return [
     |
     */
 
-    'default' => env('QUEUE_CONNECTION', 'sync'),
+    // 'default' => env('QUEUE_CONNECTION', 'sync'),
+    'default' => 'database',
 
     /*
     |--------------------------------------------------------------------------
